Spring Boot 4, Spring Shell 4

+++ src/main/java/com/internetstaff/parameterstore/adapter/in/shell/Rm.java
@@ -2,23 +2,32 @@ package com.internetstaff.parameterstore.adapter.in.shell;
 
 import com.internetstaff.parameterstore.application.port.out.ParameterStore;
 import lombok.RequiredArgsConstructor;
-import org.springframework.shell.command.annotation.Command;
-import org.springframework.shell.command.annotation.Option;
+import org.springframework.shell.core.command.annotation.Argument;
+import org.springframework.shell.core.command.annotation.Command;
+import org.springframework.shell.core.command.annotation.Option;
+import org.springframework.stereotype.Component;
 
-@Command
+@Component
 @RequiredArgsConstructor
 class Rm {
   private final ParameterStore parameterStore;
+  private final OptionPrompter promptingService;
 
   @Command(description = "Remove parameter", group = "Parameter Store")
   public String rm(
-      @Option(description = "Full path of parameter to remove", required = true) String name
+      @Option(description = "Force without prompting", shortName = 'f', defaultValue = "false") boolean force,
+      @Argument(index = 0, description = "Full path of parameter to remove") String name
   ) {
 
-    if (parameterStore.deleteParameter(name)) {
-      return "%s removed.".formatted(name);
+    if (force || promptingService.confirm("Are you sure you want to remove parameter '%s'?".formatted(name))) {
+
+      if (parameterStore.deleteParameter(name)) {
+        return "%s removed.".formatted(name);
+      } else {
+        return "Parameter not found";
+      }
     } else {
-      return "Parameter not found";
+      return "Operation cancelled";
     }
   }
 }
